requirements.txt
# LangChain & LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0

# 数据库
pymongo>=4.6.0
motor>=3.3.0

# 数据源
tushare>=1.4.0
akshare>=1.14.0

# 数据处理
pandas>=2.2.0
numpy>=1.26.0

# 配置管理
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# OCR & 图像处理
pillow>=10.2.0
pytesseract>=0.3.10

# 工具
requests>=2.31.0
httpx>=0.27.0
python-dateutil>=2.8.2

# 调度
apscheduler>=3.10.0

# 日志
loguru>=0.7.2

# Web界面
streamlit>=1.31.0

# FastAPI后端
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1.0
python-multipart>=0.0.9

# 爬虫/RSS
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# 测试
pytest>=8.0.0
pytest-cov>=4.1.0
